customised Plantis page into components
