+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import Home from './components/Home';
-import PlantisScreens from './components/PlantisScreens';
-import PlantzaScreens from './components/PlantzaScreens';
+import PlantisPage from './components/PlantisPage';
+import PlantzaPage from './components/PlantzaPage';
 import Error from './components/Error';
 import './App.css';
 
@@ -12,8 +12,8 @@ function App() {
     <BrowserRouter>
       <Switch>
         <Route path="/" component={Home} exact />
-        <Route path="/plantis" component={PlantisScreens} />
-        <Route path="/plantza" component={PlantzaScreens} />
+        <Route path="/plantis" component={PlantisPage} />
+        <Route path="/plantza" component={PlantzaPage} />
         <Route component={Error} />
       </Switch>
     </BrowserRouter>
