+++ src/components/PlantisPage.jsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import styled from 'styled-components';
+import PHeader from './plantis_components/PHeader';
+import PVideo from './plantis_components/PVideo';
+import PScreenshots from './plantis_components/PScreenshots';
+
+
+const PageWrapper = styled.div`
+  max-width: 100vw;
+  margin: 0px 50px;
+  padding: 10px 40px;
+
+  && h3 {
+    font-size: 25px;
+  }
+
+  && p {
+    font-size: 20px;
+  }
+`;
+
+const PlantisPage = () => (
+  <div style={{
+    background: 'linear-gradient(rgba(255,255,255,0), rgba(255,255,255,0)), url("https://images.pexels.com/photos/1227648/pexels-photo-1227648.jpeg?auto=compress&cs=tinysrgb&dpr=3&h=750&w=1260")',
+    backgroundPosition: 'center',
+    backgroundSize: 'cover',
+    backgroundRepeat: 'no-repeat',
+    color: 'white',
+    paddingBottom: '50px',
+  }}
+  >
+    <PageWrapper>
+      <PHeader />
+      <PVideo />
+      <PScreenshots />
+    </PageWrapper>
+  </div>
+);
+
+export default PlantisPage;
